// src/models/meal.ts
export interface Meal {
  objectId: string;
  country: string;
  courses: string[];
  cost: string;
  kitchenTimes: string[];
  city: string;
  coverImageUrl: string;
  hostId: string;
  cameraZoom: number;
  meetingPointLat: number;
  meetingPoint: string;
  phone: string;
  maxGuest: string;
  name: string;
  mealDuration: string;
  meetingPointLong: number;
  isCustomMeal: boolean;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}